+++ trixnity-messenger-compose-view/src/iosMain/kotlin/de/connect2x/trixnity/messenger/compose/view/room/timeline/element/details/PlatformPdfReader.ios.kt
@@ -15,6 +15,8 @@ import kotlinx.coroutines.NonCancellable
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
 import org.jetbrains.compose.resources.decodeToImageBitmap
+import org.koin.core.module.Module
+import org.koin.dsl.module
 import platform.CoreGraphics.CGContextFillRect
 import platform.CoreGraphics.CGContextRestoreGState
 import platform.CoreGraphics.CGContextSaveGState
@@ -32,11 +34,19 @@ import platform.UIKit.UIGraphicsGetImageFromCurrentImageContext
 import platform.UIKit.UIImage
 import platform.UIKit.UIImageJPEGRepresentation
 
-actual suspend fun getPlatformPDFReader(
-    media: PlatformMedia,
-    coroutineScope: CoroutineScope,
-    onError: (String?) -> Unit,
-): PDFReader = PlatformPDFReader(media, coroutineScope, onError).also { it.initialize() }
+actual fun getPlatformPdfReaderModule(): Module {
+    return module {
+        single<PDFReaderFactory> {
+            object : PDFReaderFactory {
+                val coroutineScope: CoroutineScope = get()
+
+                override suspend fun create(media: PlatformMedia, onError: (String?) -> Unit): PDFReader {
+                    return IosPDFReader(media, coroutineScope, onError).also { it.initialize() }
+                }
+            }
+        }
+    }
+}
 
 private val log: Logger =
     Logger("de.connect2x.trixnity.messenger.compose.view.room.timeline.element.details.PdfTimelineElementDetailsViewKt")
@@ -45,7 +55,7 @@ fun UIImage.toByteArray(compressionQuality: Double = 0.9): ByteArray? =
     UIImageJPEGRepresentation(this, compressionQuality)?.toByteArray()
 
 @OptIn(ExperimentalForeignApi::class)
-class PlatformPDFReader(
+class IosPDFReader(
     private val media: PlatformMedia,
     private val coroutineScope: CoroutineScope,
     private val onError: (String?) -> Unit,