+++ trixnity-messenger-compose-view/src/jvmMain/kotlin/de/connect2x/trixnity/messenger/compose/view/room/timeline/element/details/PlatformPdfReader.desktop.kt
@@ -16,22 +16,27 @@ import kotlinx.coroutines.withContext
 import org.apache.pdfbox.Loader
 import org.apache.pdfbox.pdmodel.PDDocument
 import org.apache.pdfbox.rendering.PDFRenderer
+import org.koin.core.module.Module
+import org.koin.dsl.module
 
 private val log: Logger =
     Logger("de.connect2x.trixnity.messenger.compose.view.room.timeline.element.details.PdfElementDetailsViewKt")
 
-actual suspend fun getPlatformPDFReader(
-    media: PlatformMedia,
-    coroutineScope: CoroutineScope,
-    onError: (String?) -> Unit,
-): PDFReader {
-    val reader = PlatformPDFReader(media, coroutineScope, onError)
-    reader.initialize()
+actual fun getPlatformPdfReaderModule(): Module {
+    return module {
+        single<PDFReaderFactory> {
+            object : PDFReaderFactory {
+                val coroutineScope: CoroutineScope = get()
 
-    return reader
+                override suspend fun create(media: PlatformMedia, onError: (String?) -> Unit): PDFReader {
+                    return DesktopPDFReader(media, coroutineScope, onError).also { it.initialize() }
+                }
+            }
+        }
+    }
 }
 
-class PlatformPDFReader(
+class DesktopPDFReader(
     val media: PlatformMedia,
     private val coroutineScope: CoroutineScope,
     val onError: (String?) -> Unit,