+++ trixnity-messenger-compose-view/src/webMain/kotlin/de/connect2x/trixnity/messenger/compose/view/room/timeline/element/details/PlatformPdfReader.web.kt
@@ -14,20 +14,26 @@ import kotlinx.coroutines.flow.filterNotNull
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
+import org.koin.core.module.Module
+import org.koin.dsl.module
 import pdfjs.GlobalWorkerOptions
 import web.blob.Blob
 
-actual suspend fun getPlatformPDFReader(
-    media: PlatformMedia,
-    coroutineScope: CoroutineScope,
-    onError: (String?) -> Unit,
-): PDFReader {
-    val reader = PDFPlatformReader(media, coroutineScope, onError)
-    reader.initialize()
-    return reader
+actual fun getPlatformPdfReaderModule(): Module {
+    return module {
+        single<PDFReaderFactory> {
+            object : PDFReaderFactory {
+                val coroutineScope: CoroutineScope = get()
+
+                override suspend fun create(media: PlatformMedia, onError: (String?) -> Unit): PDFReader {
+                    return WebPDFReader(media, coroutineScope, onError).also { it.initialize() }
+                }
+            }
+        }
+    }
 }
 
-class PDFPlatformReader(
+class WebPDFReader(
     val media: PlatformMedia,
     private val coroutineScope: CoroutineScope,
     val onError: (String?) -> Unit,